test: split out ignoreHTTPSErrors tests (#2745)
We'll get more of these soon.
